CFY-6091 Move message column and index to a mixin
